% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoviz-package.R
\docType{package}
\name{pseudoviz-package}
\alias{pseudoviz}
\alias{pseudoviz-package}
\title{pseudoviz: What the Package Does (One Line, Title Case)}
\description{
Create a hugo site from airtable.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/datasketch/pseudoviz}
  \item Report bugs at \url{https://github.com/datasketch/pseudoviz/issues}
}

}
\author{
\strong{Maintainer}: Juan Pablo Marin Diaz \email{jpmarindiaz@gmail.com}

}
\keyword{internal}
